+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Gear souls can no longer gain the unimplemented traits Chill, Warm, and Coffee Pot
 ### Fixed
 - Replaced the broken glowrose world gen configs with a new one that actually works (`glowrose.world.patchCount`) [#536]
+- Treasure Bags support updated for version 1.3.2+
 
 ## [3.7.15] - 2021-05-04
 ### Added