+++ CHANGELOG.md
@@ -3,7 +3,9 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [3.7.16] - 2021-06-15
+### Added
+- Config to blacklist glowrose world generation by dimension ID [#512]
 ### Changed
 - Gear souls can no longer gain the unimplemented traits Chill, Warm, and Coffee Pot
 ### Fixed